Stage

+++ src/TD.FE/TD.FE.TDOffice/TD.FE.TDOffice.Contracts/Dtos/MenadzmentRazduzenjeMagacinaPoOtpremnicama/MenadzmentRazduzenjeMagacinaPoOtpremnicamaPripremaDokumenataDto.cs
@@ -3,6 +3,6 @@
     public class MenadzmentRazduzenjeMagacinaPoOtpremnicamaPripremaDokumenataDto
     {
         public int UkupanBrojDokumenata { get; set; }
-        public double ZbirnaVrednostDokumenataSaCenamaNaDanasnjiDan { get; set; }
+        public decimal ZbirnaVrednostDokumenataSaCenamaNaDanasnjiDan { get; set; }
     }
 }
